refactor: move lib/postal/received_header to app/lib/received_header
