+++ app/lib/smtp_server/client.rb
@@ -366,8 +366,8 @@ module SMTPServer
       @headers = {}
       @receiving_headers = true
 
-      received_header = Postal::ReceivedHeader.generate(@credential&.server, @helo_name, @ip_address, :smtp)
-                                              .force_encoding("BINARY")
+      received_header = ReceivedHeader.generate(@credential&.server, @helo_name, @ip_address, :smtp)
+                                      .force_encoding("BINARY")
 
       @data << "Received: #{received_header}\r\n"
       @headers["received"] = [received_header]
